Sjekker på om hendelse er håndtert
